repacle anonymous classe by reusable classes
